xcode: now builds and runs for 10.4 as well
Why did people try to use NSUInteger in this? Kind of silly.

+++ X-Chat Aqua/fe-aqua/MenuMaker.mm
@@ -452,7 +452,7 @@ static MenuMaker *defaultMenuMaker;
 
 - (NSString *)stripImageFromTitle:(NSString *)title icon:(NSString **)icon
 {
-    NSUInteger length; 
+    int length;
     title = [title stringByReplacingOccurrencesOfString:@"_" withString:@""];
     length = [title length];
     if([[title substringFromIndex:length-1] isEqualToString:@"~"])
